moved MockFullTextModels to sep module
